move HistoryTests class to separate file and namespace
