+++ Builds/VisualStudio/stellar-core.vcxproj.filters
@@ -762,6 +762,9 @@
     <ClCompile Include="..\..\src\history\HistoryTests.cpp">
       <Filter>history\tests</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\history\HistoryTestsUtils.cpp">
+      <Filter>history\tests</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\herder\PendingEnvelopesTests.cpp">
       <Filter>herder\tests</Filter>
     </ClCompile>
@@ -1365,6 +1368,9 @@
       <Filter>history</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\history\HistoryManagerImpl.h">
+      <Filter>history</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\history\HistoryTestsUtils.h">
       <Filter>history</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\history\InferredQuorum.h">
