+++ src/history/HistoryTests.cpp
@@ -2,349 +2,33 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "util/asio.h"
-#include "bucket/BucketList.h"
 #include "bucket/BucketManager.h"
-#include "catchup/CatchupConfiguration.h"
-#include "catchup/CatchupWork.h"
 #include "catchup/CatchupWorkTests.h"
-#include "crypto/Hex.h"
-#include "herder/LedgerCloseData.h"
-#include "history/HistoryArchive.h"
 #include "history/HistoryManager.h"
+#include "history/HistoryTestsUtils.h"
 #include "historywork/GetHistoryArchiveStateWork.h"
 #include "historywork/GunzipFileWork.h"
 #include "historywork/GzipFileWork.h"
 #include "historywork/PutHistoryArchiveStateWork.h"
-#include "ledger/CheckpointRange.h"
 #include "ledger/LedgerManager.h"
-#include "lib/catch.hpp"
-#include "main/Application.h"
-#include "main/Config.h"
 #include "main/ExternalQueue.h"
 #include "main/PersistentState.h"
 #include "process/ProcessManager.h"
-#include "test/TestAccount.h"
 #include "test/TestUtils.h"
-#include "test/TxTests.h"
 #include "test/test.h"
 #include "util/Fs.h"
-#include "util/Logging.h"
-#include "util/NonCopyable.h"
-#include "util/Timer.h"
-#include "util/TmpDir.h"
 #include "work/WorkManager.h"
-#include "work/WorkParent.h"
 
-#include <cstdio>
-#include <fstream>
+#include <lib/catch.hpp>
 #include <lib/util/format.h>
-#include <medida/counter.h>
-#include <medida/metrics_registry.h>
-#include <random>
-#include <xdrpp/autocheck.h>
 
 using namespace stellar;
+using namespace historytestutils;
 
-namespace stellar
+TEST_CASE("next checkpoint ledger", "[history]")
 {
-using namespace txtest;
-using xdr::operator==;
-};
-
-class Configurator : NonCopyable
-{
-  public:
-    virtual Config& configure(Config& cfg, bool writable) const = 0;
-    virtual std::string
-    getArchiveDirName() const
-    {
-        return "";
-    }
-};
-
-class TmpDirConfigurator : public Configurator
-{
-    TmpDirManager mArchtmp;
-    TmpDir mDir;
-
-  public:
-    TmpDirConfigurator() : mArchtmp("archtmp"), mDir(mArchtmp.tmpDir("archive"))
-    {
-    }
-
-    std::string
-    getArchiveDirName() const override
-    {
-        return mDir.getName();
-    }
-
-    Config&
-    configure(Config& cfg, bool writable) const override
-    {
-        std::string d = mDir.getName();
-        std::string getCmd = "cp " + d + "/{0} {1}";
-        std::string putCmd = "";
-        std::string mkdirCmd = "";
-
-        if (writable)
-        {
-            putCmd = "cp {0} " + d + "/{1}";
-            mkdirCmd = "mkdir -p " + d + "/{0}";
-        }
-
-        cfg.HISTORY["test"] =
-            std::make_shared<HistoryArchive>("test", getCmd, putCmd, mkdirCmd);
-        return cfg;
-    }
-};
-
-struct CatchupMetrics
-{
-    uint64_t mHistoryArchiveStatesDownloaded;
-    uint64_t mLedgersDownloaded;
-    uint64_t mLedgersVerified;
-    uint64_t mLedgerChainsVerificationFailed;
-    uint64_t mBucketsDownloaded;
-    uint64_t mBucketsApplied;
-    uint64_t mTransactionsDownloaded;
-    uint64_t mTransactionsApplied;
-
-    CatchupMetrics()
-        : mHistoryArchiveStatesDownloaded{0}
-        , mLedgersDownloaded{0}
-        , mLedgersVerified{0}
-        , mLedgerChainsVerificationFailed{0}
-        , mBucketsDownloaded{false}
-        , mBucketsApplied{false}
-        , mTransactionsDownloaded{0}
-        , mTransactionsApplied{0}
-    {
-    }
-
-    CatchupMetrics(uint64_t historyArchiveStatesDownloaded,
-                   uint64_t ledgersDownloaded, uint64_t ledgersVerified,
-                   uint64_t ledgerChainsVerificationFailed,
-                   uint64_t bucketsDownloaded, uint64_t bucketsApplied,
-                   uint64_t transactionsDownloaded,
-                   uint64_t transactionsApplied)
-        : mHistoryArchiveStatesDownloaded{historyArchiveStatesDownloaded}
-        , mLedgersDownloaded{ledgersDownloaded}
-        , mLedgersVerified{ledgersVerified}
-        , mLedgerChainsVerificationFailed{ledgerChainsVerificationFailed}
-        , mBucketsDownloaded{bucketsDownloaded}
-        , mBucketsApplied{bucketsApplied}
-        , mTransactionsDownloaded{transactionsDownloaded}
-        , mTransactionsApplied{transactionsApplied}
-    {
-    }
-
-    friend CatchupMetrics
-    operator-(CatchupMetrics const& x, CatchupMetrics const& y)
-    {
-        return CatchupMetrics{x.mHistoryArchiveStatesDownloaded -
-                                  y.mHistoryArchiveStatesDownloaded,
-                              x.mLedgersDownloaded - y.mLedgersDownloaded,
-                              x.mLedgersVerified - y.mLedgersVerified,
-                              x.mLedgerChainsVerificationFailed -
-                                  y.mLedgerChainsVerificationFailed,
-                              x.mBucketsDownloaded - y.mBucketsDownloaded,
-                              x.mBucketsApplied - y.mBucketsApplied,
-                              x.mTransactionsDownloaded -
-                                  y.mTransactionsDownloaded,
-                              x.mTransactionsApplied - y.mTransactionsApplied};
-    }
-};
-
-struct CatchupPerformedWork
-{
-    uint64_t mHistoryArchiveStatesDownloaded;
-    uint64_t mLedgersDownloaded;
-    uint64_t mLedgersVerified;
-    uint64_t mLedgerChainsVerificationFailed;
-    bool mBucketsDownloaded;
-    bool mBucketsApplied;
-    uint64_t mTransactionsDownloaded;
-    uint64_t mTransactionsApplied;
-
-    CatchupPerformedWork(CatchupMetrics const& metrics)
-        : mHistoryArchiveStatesDownloaded{metrics
-                                              .mHistoryArchiveStatesDownloaded}
-        , mLedgersDownloaded{metrics.mLedgersDownloaded}
-        , mLedgersVerified{metrics.mLedgersVerified}
-        , mLedgerChainsVerificationFailed{metrics
-                                              .mLedgerChainsVerificationFailed}
-        , mBucketsDownloaded{metrics.mBucketsDownloaded > 0}
-        , mBucketsApplied{metrics.mBucketsApplied > 0}
-        , mTransactionsDownloaded{metrics.mTransactionsDownloaded}
-        , mTransactionsApplied{metrics.mTransactionsApplied}
-    {
-    }
-
-    CatchupPerformedWork(uint64_t historyArchiveStatesDownloaded,
-                         uint64_t ledgersDownloaded, uint64_t ledgersVerified,
-                         uint64_t ledgerChainsVerificationFailed,
-                         bool bucketsDownloaded, bool bucketsApplied,
-                         uint64_t transactionsDownloaded,
-                         uint64_t transactionsApplied)
-        : mHistoryArchiveStatesDownloaded{historyArchiveStatesDownloaded}
-        , mLedgersDownloaded{ledgersDownloaded}
-        , mLedgersVerified{ledgersVerified}
-        , mLedgerChainsVerificationFailed{ledgerChainsVerificationFailed}
-        , mBucketsDownloaded{bucketsDownloaded}
-        , mBucketsApplied{bucketsApplied}
-        , mTransactionsDownloaded{transactionsDownloaded}
-        , mTransactionsApplied{transactionsApplied}
-    {
-    }
-
-    friend bool
-    operator==(CatchupPerformedWork const& x, CatchupPerformedWork const& y)
-    {
-        if (x.mHistoryArchiveStatesDownloaded !=
-            y.mHistoryArchiveStatesDownloaded)
-        {
-            return false;
-        }
-        if (x.mLedgersDownloaded != y.mLedgersDownloaded)
-        {
-            return false;
-        }
-        if (x.mLedgersVerified != y.mLedgersVerified)
-        {
-            return false;
-        }
-        if (x.mLedgerChainsVerificationFailed !=
-            y.mLedgerChainsVerificationFailed)
-        {
-            return false;
-        }
-        if (x.mBucketsDownloaded != y.mBucketsDownloaded)
-        {
-            return false;
-        }
-        if (x.mBucketsApplied != y.mBucketsApplied)
-        {
-            return false;
-        }
-        if (x.mTransactionsDownloaded != y.mTransactionsDownloaded)
-        {
-            return false;
-        }
-        if (x.mTransactionsApplied != y.mTransactionsApplied)
-        {
-            return false;
-        }
-        return true;
-    }
-    friend bool
-    operator!=(CatchupPerformedWork const& x, CatchupPerformedWork const& y)
-    {
-        return !(x == y);
-    }
-};
-
-namespace Catch
-{
-template <>
-std::string
-toString(CatchupPerformedWork const& cm)
-{
-    return fmt::format("{}, {}, {}, {}, {}, {}, {}, {}",
-                       cm.mHistoryArchiveStatesDownloaded,
-                       cm.mLedgersDownloaded, cm.mLedgersVerified,
-                       cm.mLedgerChainsVerificationFailed,
-                       cm.mBucketsDownloaded, cm.mBucketsApplied,
-                       cm.mTransactionsDownloaded, cm.mTransactionsApplied);
-}
-}
-
-class HistoryTests
-{
-  protected:
-    VirtualClock clock;
-    std::shared_ptr<Configurator> mConfigurator;
-    Config cfg;
-    std::vector<Config> mCfgs;
-    Application::pointer mAppPtr;
-    Application& mApp;
-
-    std::default_random_engine mGenerator;
-    std::bernoulli_distribution mFlip{0.5};
-
-    std::vector<LedgerCloseData> mLedgerCloseDatas;
-
-    std::vector<uint32_t> mLedgerSeqs;
-    std::vector<uint256> mLedgerHashes;
-    std::vector<uint256> mBucketListHashes;
-    std::vector<uint256> mBucket0Hashes;
-    std::vector<uint256> mBucket1Hashes;
-
-    std::vector<int64_t> rootBalances;
-    std::vector<int64_t> aliceBalances;
-    std::vector<int64_t> bobBalances;
-    std::vector<int64_t> carolBalances;
-
-    std::vector<SequenceNumber> rootSeqs;
-    std::vector<SequenceNumber> aliceSeqs;
-    std::vector<SequenceNumber> bobSeqs;
-    std::vector<SequenceNumber> carolSeqs;
-
-  public:
-    HistoryTests(std::shared_ptr<Configurator> cg =
-                     std::make_shared<TmpDirConfigurator>())
-        : mConfigurator(cg)
-        , cfg(getTestConfig())
-        , mAppPtr(
-              createTestApplication(clock, mConfigurator->configure(cfg, true)))
-        , mApp(*mAppPtr)
-    {
-        CHECK(HistoryManager::initializeHistoryArchive(mApp, "test"));
-    }
-
-    void generateRandomLedger();
-    void generateAndPublishHistory(size_t nPublishes);
-    void generateAndPublishInitialHistory(size_t nPublishes);
-
-    Application::pointer catchupNewApplication(uint32_t initLedger,
-                                               uint32_t count, bool manual,
-                                               Config::TestDbMode dbMode,
-                                               std::string const& appName);
-
-    bool catchupApplication(uint32_t initLedger, uint32_t count, bool manual,
-                            Application::pointer app2, bool doStart = true,
-                            uint32_t gap = 0);
-
-    CatchupMetrics getCatchupMetrics(Application::pointer app);
-    CatchupPerformedWork computeCatchupPerformedWork(
-        uint32_t lastClosedLedger,
-        CatchupConfiguration const& catchupConfiguration,
-        HistoryManager const& historyManager);
-
-    bool
-    flip()
-    {
-        return mFlip(mGenerator);
-    }
-};
-
-void
-HistoryTests::generateAndPublishInitialHistory(size_t nPublishes)
-{
-    mApp.start();
-
-    auto& lm = mApp.getLedgerManager();
-
-    // At this point LCL should be 1, current ledger should be 2
-    REQUIRE(lm.getLastClosedLedgerHeader().header.ledgerSeq == 1);
-    REQUIRE(lm.getCurrentLedgerHeader().ledgerSeq == 2);
-
-    generateAndPublishHistory(nPublishes);
-}
-
-TEST_CASE_METHOD(HistoryTests, "next checkpoint ledger", "[history]")
-{
-    HistoryManager& hm = mApp.getHistoryManager();
+    HistoryTests historyTests{};
+    HistoryManager& hm = historyTests.getApp().getHistoryManager();
     CHECK(hm.nextCheckpointLedger(0) == 64);
     CHECK(hm.nextCheckpointLedger(1) == 64);
     CHECK(hm.nextCheckpointLedger(32) == 64);
@@ -360,17 +44,19 @@ TEST_CASE_METHOD(HistoryTests, "next checkpoint ledger", "[history]")
     CHECK(hm.nextCheckpointLedger(130) == 192);
 }
 
-TEST_CASE_METHOD(HistoryTests, "HistoryManager::compress", "[history]")
+TEST_CASE("HistoryManager::compress", "[history]")
 {
+    HistoryTests historyTests{};
+
     std::string s = "hello there";
-    HistoryManager& hm = mApp.getHistoryManager();
+    HistoryManager& hm = historyTests.getApp().getHistoryManager();
     std::string fname = hm.localFilename("compressme");
     {
         std::ofstream out(fname, std::ofstream::binary);
         out.write(s.data(), s.size());
     }
     std::string compressed = fname + ".gz";
-    auto& wm = mApp.getWorkManager();
+    auto& wm = historyTests.getApp().getWorkManager();
     auto g = wm.executeWork<GzipFileWork>(true, fname);
     REQUIRE(g->getState() == Work::WORK_SUCCESS);
     REQUIRE(!fs::exists(fname));
@@ -382,18 +68,20 @@ TEST_CASE_METHOD(HistoryTests, "HistoryManager::compress", "[history]")
     REQUIRE(!fs::exists(compressed));
 }
 
-TEST_CASE_METHOD(HistoryTests, "HistoryArchiveState::get_put", "[history]")
+TEST_CASE("HistoryArchiveState::get_put", "[history]")
 {
+    HistoryTests historyTests{};
+
     HistoryArchiveState has;
     has.currentLedger = 0x1234;
 
-    auto i = mApp.getConfig().HISTORY.find("test");
-    REQUIRE(i != mApp.getConfig().HISTORY.end());
+    auto i = historyTests.getApp().getConfig().HISTORY.find("test");
+    REQUIRE(i != historyTests.getApp().getConfig().HISTORY.end());
     auto archive = i->second;
 
     has.resolveAllFutures();
 
-    auto& wm = mApp.getWorkManager();
+    auto& wm = historyTests.getApp().getWorkManager();
     auto put = wm.executeWork<PutHistoryArchiveStateWork>(true, has, archive);
     REQUIRE(put->getState() == Work::WORK_SUCCESS);
 
@@ -405,458 +93,11 @@ TEST_CASE_METHOD(HistoryTests, "HistoryArchiveState::get_put", "[history]")
     REQUIRE(has2.currentLedger == 0x1234);
 }
 
-void
-HistoryTests::generateRandomLedger()
+TEST_CASE("History publish", "[history]")
 {
-    auto& lm = mApp.getLedgerManager();
-    TxSetFramePtr txSet =
-        std::make_shared<TxSetFrame>(lm.getLastClosedLedgerHeader().hash);
+    HistoryTests historyTests{};
 
-    uint32_t ledgerSeq = lm.getLedgerNum();
-    uint64_t minBalance = lm.getMinBalance(5);
-    uint64_t big = minBalance + ledgerSeq;
-    uint64_t small = 100 + ledgerSeq;
-    uint64_t closeTime = 60 * 5 * ledgerSeq;
-
-    auto root = TestAccount{mApp, getRoot(mApp.getNetworkID())};
-    auto alice = TestAccount{mApp, getAccount("alice")};
-    auto bob = TestAccount{mApp, getAccount("bob")};
-    auto carol = TestAccount{mApp, getAccount("carol")};
-
-    // Root sends to alice every tx, bob every other tx, carol every 4rd tx.
-    txSet->add(root.tx({createAccount(alice, big)}));
-    txSet->add(root.tx({createAccount(bob, big)}));
-    txSet->add(root.tx({createAccount(carol, big)}));
-    txSet->add(root.tx({payment(alice, big)}));
-    txSet->add(root.tx({payment(bob, big)}));
-    txSet->add(root.tx({payment(carol, big)}));
-
-    // They all randomly send a little to one another every ledger after #4
-    if (ledgerSeq > 4)
-    {
-        if (flip())
-            txSet->add(alice.tx({payment(bob, small)}));
-        if (flip())
-            txSet->add(alice.tx({payment(carol, small)}));
-
-        if (flip())
-            txSet->add(bob.tx({payment(alice, small)}));
-        if (flip())
-            txSet->add(bob.tx({payment(carol, small)}));
-
-        if (flip())
-            txSet->add(carol.tx({payment(alice, small)}));
-        if (flip())
-            txSet->add(carol.tx({payment(bob, small)}));
-    }
-
-    // Provoke sortForHash and hash-caching:
-    txSet->getContentsHash();
-
-    CLOG(DEBUG, "History") << "Closing synthetic ledger " << ledgerSeq
-                           << " with " << txSet->size() << " txs (txhash:"
-                           << hexAbbrev(txSet->getContentsHash()) << ")";
-
-    StellarValue sv(txSet->getContentsHash(), closeTime, emptyUpgradeSteps, 0);
-    mLedgerCloseDatas.emplace_back(ledgerSeq, txSet, sv);
-    lm.closeLedger(mLedgerCloseDatas.back());
-
-    mLedgerSeqs.push_back(lm.getLastClosedLedgerHeader().header.ledgerSeq);
-    mLedgerHashes.push_back(lm.getLastClosedLedgerHeader().hash);
-    mBucketListHashes.push_back(
-        lm.getLastClosedLedgerHeader().header.bucketListHash);
-    mBucket0Hashes.push_back(mApp.getBucketManager()
-                                 .getBucketList()
-                                 .getLevel(0)
-                                 .getCurr()
-                                 ->getHash());
-    mBucket1Hashes.push_back(mApp.getBucketManager()
-                                 .getBucketList()
-                                 .getLevel(2)
-                                 .getCurr()
-                                 ->getHash());
-
-    rootBalances.push_back(root.getBalance());
-    aliceBalances.push_back(alice.getBalance());
-    bobBalances.push_back(bob.getBalance());
-    carolBalances.push_back(carol.getBalance());
-
-    rootSeqs.push_back(root.loadSequenceNumber());
-    aliceSeqs.push_back(alice.loadSequenceNumber());
-    bobSeqs.push_back(bob.loadSequenceNumber());
-    carolSeqs.push_back(carol.loadSequenceNumber());
-}
-
-void
-HistoryTests::generateAndPublishHistory(size_t nPublishes)
-{
-    auto& lm = mApp.getLedgerManager();
-    auto& hm = mApp.getHistoryManager();
-
-    size_t publishSuccesses = hm.getPublishSuccessCount();
-    SequenceNumber ledgerSeq = lm.getCurrentLedgerHeader().ledgerSeq;
-
-    while (hm.getPublishSuccessCount() < (publishSuccesses + nPublishes))
-    {
-        uint64_t queueCount = hm.getPublishQueueCount();
-        while (hm.getPublishQueueCount() == queueCount)
-        {
-            generateRandomLedger();
-            ++ledgerSeq;
-        }
-
-        REQUIRE(lm.getCurrentLedgerHeader().ledgerSeq == ledgerSeq);
-
-        // Advance until we've published (or failed to!)
-        while (hm.getPublishSuccessCount() < hm.getPublishQueueCount())
-        {
-            REQUIRE(hm.getPublishFailureCount() == 0);
-            mApp.getClock().crank(true);
-        }
-    }
-
-    REQUIRE(hm.getPublishFailureCount() == 0);
-    REQUIRE(hm.getPublishSuccessCount() == publishSuccesses + nPublishes);
-    REQUIRE(lm.getLedgerNum() ==
-            ((publishSuccesses + nPublishes) * hm.getCheckpointFrequency()));
-}
-
-Application::pointer
-HistoryTests::catchupNewApplication(uint32_t initLedger, uint32_t count,
-                                    bool manual, Config::TestDbMode dbMode,
-                                    std::string const& appName)
-{
-
-    CLOG(INFO, "History") << "****";
-    CLOG(INFO, "History") << "**** Beginning catchup test for app '" << appName
-                          << "'";
-    CLOG(INFO, "History") << "****";
-
-    mCfgs.emplace_back(
-        getTestConfig(static_cast<int>(mCfgs.size()) + 1, dbMode));
-    mCfgs.back().CATCHUP_COMPLETE =
-        count == std::numeric_limits<uint32_t>::max();
-    if (count != std::numeric_limits<uint32_t>::max())
-    {
-        mCfgs.back().CATCHUP_RECENT = count;
-    }
-    Application::pointer app2 = createTestApplication(
-        clock, mConfigurator->configure(mCfgs.back(), false));
-
-    app2->start();
-    CHECK(catchupApplication(initLedger, count, manual, app2) == true);
-    return app2;
-}
-
-bool
-HistoryTests::catchupApplication(uint32_t initLedger, uint32_t count,
-                                 bool manual, Application::pointer app2,
-                                 bool doStart, uint32_t gap)
-{
-    auto startCatchupMetrics = getCatchupMetrics(app2);
-
-    auto root = TestAccount{*app2, getRoot(mApp.getNetworkID())};
-    auto alice = TestAccount{*app2, getAccount("alice")};
-    auto bob = TestAccount{*app2, getAccount("bob")};
-    auto carol = TestAccount{*app2, getAccount("carol")};
-
-    auto& lm = app2->getLedgerManager();
-    if (doStart)
-    {
-        // Normally Herder calls LedgerManager.externalizeValue(initLedger + 1)
-        // and this _triggers_ catchup within the LM. However, we do this
-        // out-of-order because we want to control the catchup mode rather than
-        // let the LM pick it, and because we want to simulate a 1-ledger skew
-        // between the publishing side and the catchup side so that the catchup
-        // has "heard" exactly 1 consensus LedgerCloseData broadcast after the
-        // event that triggered its catchup to begin.
-        //
-        // For example: we want initLedger to be (say) 191-or-less, so that it
-        // catches up using block 3, but we want the publisher to advance past
-        // 192 (the first entry in block 4) and externalize that value, so that
-        // the catchup can see a {192}.prevHash to knit up block 3 against.
-
-        CLOG(INFO, "History")
-            << "force-starting catchup at initLedger=" << initLedger;
-
-        lm.startCatchUp({initLedger, count}, manual);
-    }
-
-    // Push publishing side forward one-ledger into a history block if it's
-    // sitting on the boundary of it. This will ensure there's something
-    // externalizable to knit-up with on the catchup side.
-    if (mApp.getHistoryManager().nextCheckpointLedger(
-            mApp.getLedgerManager().getLastClosedLedgerNum()) ==
-        mApp.getLedgerManager().getLedgerNum())
-    {
-        CLOG(INFO, "History")
-            << "force-publishing first ledger in next history block, ledger="
-            << mApp.getLedgerManager().getLedgerNum();
-        generateRandomLedger();
-    }
-
-    // Externalize (to the catchup LM) the range of ledgers between initLedger
-    // and as near as we can get to the first ledger of the block after
-    // initLedger (inclusive), so that there's something to knit-up with. Do not
-    // externalize anything we haven't yet published, of course.
-    if (!manual)
-    {
-        uint32_t nextBlockStart =
-            mApp.getHistoryManager().nextCheckpointLedger(initLedger);
-        for (uint32_t n = initLedger + 1; n <= nextBlockStart; ++n)
-        {
-            // Remember the vectors count from 2, not 0.
-            if (n - 2 >= mLedgerCloseDatas.size())
-            {
-                break;
-            }
-            if (n == gap)
-            {
-                CLOG(INFO, "History")
-                    << "simulating LedgerClose transmit gap at ledger " << n;
-            }
-            else
-            {
-                // Remember the vectors count from 2, not 0.
-                auto const& lcd = mLedgerCloseDatas.at(n - 2);
-                CLOG(INFO, "History")
-                    << "force-externalizing LedgerCloseData for " << n
-                    << " has txhash:"
-                    << hexAbbrev(lcd.getTxSet()->getContentsHash());
-                lm.valueExternalized(lcd);
-            }
-        }
-    }
-
-    uint32_t lastLedger = lm.getLastClosedLedgerNum();
-    auto catchupConfiguration = CatchupConfiguration(initLedger, count);
-
-    REQUIRE(!app2->getClock().getIOService().stopped());
-
-    while (!app2->getWorkManager().allChildrenDone())
-    {
-        app2->getClock().crank(false);
-    }
-
-    if (app2->getLedgerManager().getState() != LedgerManager::LM_SYNCED_STATE)
-    {
-        return false;
-    }
-
-    auto endCatchupMetrics = getCatchupMetrics(app2);
-    auto catchupPerformedWork =
-        CatchupPerformedWork{endCatchupMetrics - startCatchupMetrics};
-
-    REQUIRE(catchupPerformedWork ==
-            computeCatchupPerformedWork(lastLedger, catchupConfiguration,
-                                        app2->getHistoryManager()));
-
-    uint32_t nextLedger = lm.getLedgerNum();
-
-    CLOG(INFO, "History") << "Caught up: lastLedger = " << lastLedger;
-    CLOG(INFO, "History") << "Caught up: initLedger = " << initLedger;
-    CLOG(INFO, "History") << "Caught up: nextLedger = " << nextLedger;
-    CLOG(INFO, "History") << "Caught up: published range is "
-                          << mLedgerSeqs.size() << " ledgers, covering "
-                          << "[" << mLedgerSeqs.front() << ", "
-                          << mLedgerSeqs.back() << "]";
-
-    // Assuming we caught up to nextLedger 128 (say), LCL will be 127, so we
-    // must subtract 1.
-    //
-    // The local history vectors are built starting from ledger 2 (put at
-    // vector-entry 0), so
-    // to access slot 127 we must subtract 2 more.
-    //
-    // So cumulatively: we want to probe local history slot i = nextLedger - 3.
-
-    REQUIRE(nextLedger != 0);
-    if (nextLedger >= 3)
-    {
-        size_t i = nextLedger - 3;
-
-        auto wantSeq = mLedgerSeqs.at(i);
-        auto wantHash = mLedgerHashes.at(i);
-        auto wantBucketListHash = mBucketListHashes.at(i);
-        auto wantBucket0Hash = mBucket0Hashes.at(i);
-        auto wantBucket1Hash = mBucket1Hashes.at(i);
-
-        auto haveSeq = lm.getLastClosedLedgerHeader().header.ledgerSeq;
-        auto haveHash = lm.getLastClosedLedgerHeader().hash;
-        auto haveBucketListHash =
-            lm.getLastClosedLedgerHeader().header.bucketListHash;
-        auto haveBucket0Hash = app2->getBucketManager()
-                                   .getBucketList()
-                                   .getLevel(0)
-                                   .getCurr()
-                                   ->getHash();
-        auto haveBucket1Hash = app2->getBucketManager()
-                                   .getBucketList()
-                                   .getLevel(2)
-                                   .getCurr()
-                                   ->getHash();
-
-        CLOG(INFO, "History")
-            << "Caught up: want Seq[" << i << "] = " << wantSeq;
-        CLOG(INFO, "History")
-            << "Caught up: have Seq[" << i << "] = " << haveSeq;
-
-        CLOG(INFO, "History")
-            << "Caught up: want Hash[" << i << "] = " << hexAbbrev(wantHash);
-        CLOG(INFO, "History")
-            << "Caught up: have Hash[" << i << "] = " << hexAbbrev(haveHash);
-
-        CLOG(INFO, "History") << "Caught up: want BucketListHash[" << i
-                              << "] = " << hexAbbrev(wantBucketListHash);
-        CLOG(INFO, "History") << "Caught up: have BucketListHash[" << i
-                              << "] = " << hexAbbrev(haveBucketListHash);
-
-        CLOG(INFO, "History") << "Caught up: want Bucket0Hash[" << i
-                              << "] = " << hexAbbrev(wantBucket0Hash);
-        CLOG(INFO, "History") << "Caught up: have Bucket0Hash[" << i
-                              << "] = " << hexAbbrev(haveBucket0Hash);
-
-        CLOG(INFO, "History") << "Caught up: want Bucket1Hash[" << i
-                              << "] = " << hexAbbrev(wantBucket1Hash);
-        CLOG(INFO, "History") << "Caught up: have Bucket1Hash[" << i
-                              << "] = " << hexAbbrev(haveBucket1Hash);
-
-        CHECK(nextLedger == haveSeq + 1);
-        CHECK(wantSeq == haveSeq);
-        CHECK(wantBucketListHash == haveBucketListHash);
-        CHECK(wantHash == haveHash);
-
-        CHECK(app2->getBucketManager().getBucketByHash(wantBucket0Hash));
-        CHECK(app2->getBucketManager().getBucketByHash(wantBucket1Hash));
-        CHECK(wantBucket0Hash == haveBucket0Hash);
-        CHECK(wantBucket1Hash == haveBucket1Hash);
-
-        auto haveRootBalance = rootBalances.at(i);
-        auto haveAliceBalance = aliceBalances.at(i);
-        auto haveBobBalance = bobBalances.at(i);
-        auto haveCarolBalance = carolBalances.at(i);
-
-        auto haveRootSeq = rootSeqs.at(i);
-        auto haveAliceSeq = aliceSeqs.at(i);
-        auto haveBobSeq = bobSeqs.at(i);
-        auto haveCarolSeq = carolSeqs.at(i);
-
-        auto wantRootBalance = root.getBalance();
-        auto wantAliceBalance = alice.getBalance();
-        auto wantBobBalance = bob.getBalance();
-        auto wantCarolBalance = carol.getBalance();
-
-        auto wantRootSeq = root.loadSequenceNumber();
-        auto wantAliceSeq = alice.loadSequenceNumber();
-        auto wantBobSeq = bob.loadSequenceNumber();
-        auto wantCarolSeq = carol.loadSequenceNumber();
-
-        CHECK(haveRootBalance == wantRootBalance);
-        CHECK(haveAliceBalance == wantAliceBalance);
-        CHECK(haveBobBalance == wantBobBalance);
-        CHECK(haveCarolBalance == wantCarolBalance);
-
-        CHECK(haveRootSeq == wantRootSeq);
-        CHECK(haveAliceSeq == wantAliceSeq);
-        CHECK(haveBobSeq == wantBobSeq);
-        CHECK(haveCarolSeq == wantCarolSeq);
-    }
-
-    mApp.getLedgerManager().checkDbState();
-    return true;
-}
-
-CatchupMetrics
-HistoryTests::getCatchupMetrics(Application::pointer app)
-{
-    auto& getHistoryArchiveStateSuccess = app->getMetrics().NewMeter(
-        {"history", "download-history-archive-state", "success"}, "event");
-    auto historyArchiveStatesDownloaded = getHistoryArchiveStateSuccess.count();
-
-    auto& downloadLedgersCached = app->getMetrics().NewMeter(
-        {"history", "download-ledger", "cached"}, "event");
-    auto& downloadLedgersSuccess = app->getMetrics().NewMeter(
-        {"history", "download-ledger", "success"}, "event");
-
-    auto ledgersDownloaded =
-        downloadLedgersSuccess.count() + downloadLedgersCached.count();
-
-    auto& verifyLedgerSuccess = app->getMetrics().NewMeter(
-        {"history", "verify-ledger", "success"}, "event");
-    auto& verifyLedgerChainFailure = app->getMetrics().NewMeter(
-        {"history", "verify-ledger-chain", "failure"}, "event");
-
-    auto ledgersVerified = verifyLedgerSuccess.count();
-    auto ledgerChainsVerificationFailed = verifyLedgerChainFailure.count();
-
-    auto& downloadBucketSuccess = app->getMetrics().NewMeter(
-        {"history", "download-bucket", "success"}, "event");
-
-    auto bucketsDownloaded = downloadBucketSuccess.count();
-
-    auto& bucketApplySuccess = app->getMetrics().NewMeter(
-        {"history", "bucket-apply", "success"}, "event");
-
-    auto bucketsApplied = bucketApplySuccess.count();
-
-    auto& downloadTransactionsCached = app->getMetrics().NewMeter(
-        {"history", "download-transactions", " cached "}, "event");
-    auto& downloadTransactionsSuccess = app->getMetrics().NewMeter(
-        {"history", "download-transactions", "success"}, "event");
-
-    auto transactionsDownloaded = downloadTransactionsSuccess.count() +
-                                  downloadTransactionsCached.count();
-
-    auto& applyLedgerSuccess = app->getMetrics().NewMeter(
-        {"history", "apply-ledger", "success"}, "event");
-
-    auto transactionsApplied = applyLedgerSuccess.count();
-
-    return CatchupMetrics{
-        historyArchiveStatesDownloaded, ledgersDownloaded,  ledgersVerified,
-        ledgerChainsVerificationFailed, bucketsDownloaded,  bucketsApplied,
-        transactionsDownloaded,         transactionsApplied};
-}
-
-CatchupPerformedWork
-HistoryTests::computeCatchupPerformedWork(
-    uint32_t lastClosedLedger, CatchupConfiguration const& catchupConfiguration,
-    HistoryManager const& historyManager)
-{
-    auto catchupRange = CatchupWork::makeCatchupRange(
-        lastClosedLedger, catchupConfiguration, historyManager);
-    auto checkpointRange = CheckpointRange{catchupRange.first, historyManager};
-
-    uint32_t historyArchiveStatesDownloaded = 1;
-    if (catchupRange.second &&
-        checkpointRange.first() != checkpointRange.last())
-    {
-        historyArchiveStatesDownloaded++;
-    }
-
-    auto filesDownloaded = checkpointRange.count();
-    auto firstVerifiedLedger = std::max(
-        LedgerManager::GENESIS_LEDGER_SEQ,
-        checkpointRange.first() + 1 - historyManager.getCheckpointFrequency());
-    auto ledgersVerified =
-        catchupConfiguration.toLedger() - firstVerifiedLedger + 1;
-    auto transactionsApplied =
-        catchupRange.second
-            ? catchupConfiguration.toLedger() - checkpointRange.first()
-            : catchupConfiguration.toLedger() - lastClosedLedger;
-    return {historyArchiveStatesDownloaded,
-            filesDownloaded,
-            ledgersVerified,
-            0,
-            catchupRange.second,
-            catchupRange.second,
-            filesDownloaded,
-            transactionsApplied};
-}
-
-TEST_CASE_METHOD(HistoryTests, "History publish", "[history]")
-{
-    generateAndPublishInitialHistory(1);
+    historyTests.generateAndPublishInitialHistory(1);
 }
 
 static std::string
@@ -891,12 +132,14 @@ dbModeName(Config::TestDbMode mode)
     }
 }
 
-TEST_CASE_METHOD(HistoryTests, "Full history catchup",
-                 "[history][historycatchup]")
+TEST_CASE("Full history catchup", "[history][historycatchup]")
 {
-    generateAndPublishInitialHistory(3);
+    HistoryTests historyTests{};
 
-    uint32_t initLedger = mApp.getLedgerManager().getLastClosedLedgerNum();
+    historyTests.generateAndPublishInitialHistory(3);
+
+    uint32_t initLedger =
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
 
     std::vector<Application::pointer> apps;
 
@@ -914,25 +157,26 @@ TEST_CASE_METHOD(HistoryTests, "Full history catchup",
     {
         for (auto count : counts)
         {
-            auto a = catchupNewApplication(initLedger, count, false, dbMode,
-                                           std::string("full, ") +
-                                               resumeModeName(count) + ", " +
-                                               dbModeName(dbMode));
+            auto a = historyTests.catchupNewApplication(
+                initLedger, count, false, dbMode,
+                std::string("full, ") + resumeModeName(count) + ", " +
+                    dbModeName(dbMode));
             apps.push_back(a);
         }
     }
 }
 
-TEST_CASE_METHOD(HistoryTests, "History publish queueing",
-                 "[history][historydelay][historycatchup]")
+TEST_CASE("History publish queueing", "[history][historydelay][historycatchup]")
 {
-    generateAndPublishInitialHistory(1);
+    HistoryTests historyTests{};
 
-    auto& hm = mApp.getHistoryManager();
+    historyTests.generateAndPublishInitialHistory(1);
+
+    auto& hm = historyTests.getApp().getHistoryManager();
 
     while (hm.getPublishQueueCount() < 4)
     {
-        generateRandomLedger();
+        historyTests.generateRandomLedger();
     }
     CLOG(INFO, "History") << "publish-delay count: "
                           << hm.getPublishDelayCount();
@@ -940,27 +184,29 @@ TEST_CASE_METHOD(HistoryTests, "History publish queueing",
     while (hm.getPublishSuccessCount() < hm.getPublishQueueCount())
     {
         CHECK(hm.getPublishFailureCount() == 0);
-        mApp.getClock().crank(true);
+        historyTests.getApp().getClock().crank(true);
     }
 
-    auto initLedger = mApp.getLedgerManager().getLastClosedLedgerNum();
-    auto app2 =
-        catchupNewApplication(initLedger, std::numeric_limits<uint32_t>::max(),
-                              false, Config::TESTDB_IN_MEMORY_SQLITE,
-                              std::string("Catchup to delayed history"));
+    auto initLedger =
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
+    auto app2 = historyTests.catchupNewApplication(
+        initLedger, std::numeric_limits<uint32_t>::max(), false,
+        Config::TESTDB_IN_MEMORY_SQLITE,
+        std::string("Catchup to delayed history"));
     CHECK(app2->getLedgerManager().getLedgerNum() ==
-          mApp.getLedgerManager().getLedgerNum());
+          historyTests.getApp().getLedgerManager().getLedgerNum());
 }
 
-TEST_CASE_METHOD(HistoryTests, "History prefix catchup",
-                 "[history][historycatchup][prefixcatchup]")
+TEST_CASE("History prefix catchup", "[history][historycatchup][prefixcatchup]")
 {
-    generateAndPublishInitialHistory(3);
+    HistoryTests historyTests{};
+
+    historyTests.generateAndPublishInitialHistory(3);
     std::vector<Application::pointer> apps;
 
     // First attempt catchup to 10, prefix of 64. Should round up to 64.
     // Should replay the 64th (since it gets externalized) and land on 65.
-    auto a = catchupNewApplication(
+    auto a = historyTests.catchupNewApplication(
         10, std::numeric_limits<uint32_t>::max(), false,
         Config::TESTDB_IN_MEMORY_SQLITE,
         std::string("Catchup to prefix of published history"));
@@ -970,7 +216,7 @@ TEST_CASE_METHOD(HistoryTests, "History prefix catchup",
 
     // Then attempt catchup to 74, prefix of 128. Should round up to 128.
     // Should replay the 64th (since it gets externalized) and land on 129.
-    a = catchupNewApplication(
+    a = historyTests.catchupNewApplication(
         freq + 10, std::numeric_limits<uint32_t>::max(), false,
         Config::TESTDB_IN_MEMORY_SQLITE,
         std::string("Catchup to second prefix of published history"));
@@ -978,27 +224,29 @@ TEST_CASE_METHOD(HistoryTests, "History prefix catchup",
     CHECK(apps.back()->getLedgerManager().getLedgerNum() == 2 * freq + 1);
 }
 
-TEST_CASE_METHOD(HistoryTests, "Publish/catchup alternation, with stall",
-                 "[history][historycatchup][catchupalternation]")
+TEST_CASE("Publish/catchup alternation, with stall",
+          "[history][historycatchup][catchupalternation]")
 {
+    HistoryTests historyTests{};
+
     // Publish in app, catch up in app2 and app3.
     // App2 will catch up using CATCHUP_COMPLETE, app3 will use
     // CATCHUP_MINIMAL.
-    generateAndPublishInitialHistory(3);
+    historyTests.generateAndPublishInitialHistory(3);
 
     Application::pointer app2, app3;
 
-    auto& lm = mApp.getLedgerManager();
+    auto& lm = historyTests.getApp().getLedgerManager();
 
     uint32_t initLedger = lm.getLastClosedLedgerNum();
 
-    app2 = catchupNewApplication(
+    app2 = historyTests.catchupNewApplication(
         initLedger, std::numeric_limits<uint32_t>::max(), false,
         Config::TESTDB_IN_MEMORY_SQLITE, std::string("app2"));
 
-    app3 = catchupNewApplication(initLedger, 0, false,
-                                 Config::TESTDB_IN_MEMORY_SQLITE,
-                                 std::string("app3"));
+    app3 = historyTests.catchupNewApplication(initLedger, 0, false,
+                                              Config::TESTDB_IN_MEMORY_SQLITE,
+                                              std::string("app3"));
 
     CHECK(app2->getLedgerManager().getLedgerNum() == lm.getLedgerNum());
     CHECK(app3->getLedgerManager().getLedgerNum() == lm.getLedgerNum());
@@ -1006,13 +254,13 @@ TEST_CASE_METHOD(HistoryTests, "Publish/catchup alternation, with stall",
     for (size_t i = 1; i < 4; ++i)
     {
         // Now alternate between publishing new stuff and catching up to it.
-        generateAndPublishHistory(i);
+        historyTests.generateAndPublishHistory(i);
 
         initLedger = lm.getLastClosedLedgerNum();
 
-        catchupApplication(initLedger, std::numeric_limits<uint32_t>::max(),
-                           false, app2);
-        catchupApplication(initLedger, 0, false, app3);
+        historyTests.catchupApplication(
+            initLedger, std::numeric_limits<uint32_t>::max(), false, app2);
+        historyTests.catchupApplication(initLedger, 0, false, app3);
 
         CHECK(app2->getLedgerManager().getLedgerNum() == lm.getLedgerNum());
         CHECK(app3->getLedgerManager().getLedgerNum() == lm.getLedgerNum());
@@ -1027,9 +275,9 @@ TEST_CASE_METHOD(HistoryTests, "Publish/catchup alternation, with stall",
     // Finally, publish a little more history than the last publish-point
     // but not enough to get to the _next_ publish-point:
 
-    generateRandomLedger();
-    generateRandomLedger();
-    generateRandomLedger();
+    historyTests.generateRandomLedger();
+    historyTests.generateRandomLedger();
+    historyTests.generateRandomLedger();
 
     // Attempting to catch up here should _stall_. We evaluate stalling
     // by providing 30 cranks of the event loop and assuming that failure
@@ -1037,72 +285,80 @@ TEST_CASE_METHOD(HistoryTests, "Publish/catchup alternation, with stall",
 
     initLedger = lm.getLastClosedLedgerNum();
 
-    REQUIRE(!catchupApplication(
+    REQUIRE(!historyTests.catchupApplication(
         initLedger, std::numeric_limits<uint32_t>::max(), false, app2));
-    REQUIRE(!catchupApplication(initLedger, 0, false, app3));
+    REQUIRE(!historyTests.catchupApplication(initLedger, 0, false, app3));
 
     // Now complete this publish cycle and confirm that the stalled apps
     // will catch up.
-    generateAndPublishHistory(1);
-    REQUIRE(catchupApplication(initLedger, std::numeric_limits<uint32_t>::max(),
-                               false, app2, false));
-    REQUIRE(catchupApplication(initLedger, 0, false, app3, false));
+    historyTests.generateAndPublishHistory(1);
+    REQUIRE(historyTests.catchupApplication(
+        initLedger, std::numeric_limits<uint32_t>::max(), false, app2, false));
+    REQUIRE(historyTests.catchupApplication(initLedger, 0, false, app3, false));
 }
 
-TEST_CASE_METHOD(HistoryTests, "Repair missing buckets via history",
-                 "[history][historybucketrepair]")
+TEST_CASE("Repair missing buckets via history",
+          "[history][historybucketrepair]")
 {
-    generateAndPublishInitialHistory(1);
+    HistoryTests historyTests{};
+
+    historyTests.generateAndPublishInitialHistory(1);
 
     // Forcibly resolve any merges in progress, so we have a calm state to
     // repair;
     // NB: we cannot repair lost buckets from merges-in-progress, as they're
     // not
     // necessarily _published_ anywhere.
-    HistoryArchiveState has(mApp.getLedgerManager().getLastClosedLedgerNum(),
-                            mApp.getBucketManager().getBucketList());
+    HistoryArchiveState has(
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum(),
+        historyTests.getApp().getBucketManager().getBucketList());
     has.resolveAllFutures();
     auto state = has.toString();
 
     auto cfg2 = getTestConfig(1);
     cfg2.BUCKET_DIR_PATH += "2";
-    auto app2 =
-        createTestApplication(clock, mConfigurator->configure(cfg2, false));
+    auto app2 = createTestApplication(
+        historyTests.getClock(),
+        historyTests.getConfigurator().configure(cfg2, false));
     app2->getPersistentState().setState(PersistentState::kHistoryArchiveState,
                                         state);
 
     app2->start();
 
-    auto hash1 = mApp.getBucketManager().getBucketList().getHash();
+    auto hash1 =
+        historyTests.getApp().getBucketManager().getBucketList().getHash();
     auto hash2 = app2->getBucketManager().getBucketList().getHash();
     CHECK(hash1 == hash2);
 }
 
-TEST_CASE_METHOD(HistoryTests, "Repair missing buckets fails",
-                 "[history][historybucketrepair]")
+TEST_CASE("Repair missing buckets fails", "[history][historybucketrepair]")
 {
-    generateAndPublishInitialHistory(1);
+    HistoryTests historyTests{};
+
+    historyTests.generateAndPublishInitialHistory(1);
 
     // Forcibly resolve any merges in progress, so we have a calm state to
     // repair;
     // NB: we cannot repair lost buckets from merges-in-progress, as they're
     // not
     // necessarily _published_ anywhere.
-    HistoryArchiveState has(mApp.getLedgerManager().getLastClosedLedgerNum(),
-                            mApp.getBucketManager().getBucketList());
+    HistoryArchiveState has(
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum(),
+        historyTests.getApp().getBucketManager().getBucketList());
     has.resolveAllFutures();
     auto state = has.toString();
 
     // Delete buckets from the archive before proceding.
     // This means startup will fail.
-    auto dir = mConfigurator->getArchiveDirName();
+    auto dir = historyTests.getConfigurator().getArchiveDirName();
     REQUIRE(!dir.empty());
     fs::deltree(dir + "/bucket");
 
     auto cfg2 = getTestConfig(1);
     cfg2.BUCKET_DIR_PATH += "2";
-    auto app2 =
-        createTestApplication(clock, mConfigurator->configure(cfg2, false));
+    auto app2 = createTestApplication(
+        historyTests.getClock(),
+        historyTests.getConfigurator().configure(cfg2, false));
     app2->getPersistentState().setState(PersistentState::kHistoryArchiveState,
                                         state);
 
@@ -1123,48 +379,16 @@ TEST_CASE_METHOD(HistoryTests, "Repair missing buckets fails",
     }
 }
 
-class S3Configurator : public Configurator
+TEST_CASE("Publish/catchup via s3", "[hide][s3]")
 {
-  public:
-    Config&
-    configure(Config& cfg, bool writable) const override
-    {
-        char const* s3bucket = getenv("S3BUCKET");
-        if (!s3bucket)
-        {
-            throw std::runtime_error("s3 test requires S3BUCKET env var");
-        }
-        std::string s3b(s3bucket);
-        if (s3b.find("s3://") != 0)
-        {
-            s3b = std::string("s3://") + s3b;
-        }
-        std::string getCmd = "aws s3 cp " + s3b + "/{0} {1}";
-        std::string putCmd = "";
-        std::string mkdirCmd = "";
-        if (writable)
-        {
-            putCmd = "aws s3 cp {0} " + s3b + "/{1}";
-        }
-        cfg.HISTORY["test"] =
-            std::make_shared<HistoryArchive>("test", getCmd, putCmd, mkdirCmd);
-        return cfg;
-    }
-};
+    HistoryTests historyTests{std::make_shared<S3Configurator>()};
 
-class S3HistoryTests : public HistoryTests
-{
-  public:
-    S3HistoryTests() : HistoryTests(std::make_shared<S3Configurator>())
-    {
-    }
-};
-
-TEST_CASE_METHOD(S3HistoryTests, "Publish/catchup via s3", "[hide][s3]")
-{
-    generateAndPublishInitialHistory(3);
-    auto app2 = catchupNewApplication(
-        mApp.getLedgerManager().getCurrentLedgerHeader().ledgerSeq,
+    historyTests.generateAndPublishInitialHistory(3);
+    auto app2 = historyTests.catchupNewApplication(
+        historyTests.getApp()
+            .getLedgerManager()
+            .getCurrentLedgerHeader()
+            .ledgerSeq,
         std::numeric_limits<uint32_t>::max(), false,
         Config::TESTDB_IN_MEMORY_SQLITE, "s3");
 }
@@ -1239,38 +463,40 @@ TEST_CASE("persist publish queue", "[history]")
 // The idea with this test is that we join a network and somehow get a gap
 // in the SCP voting sequence while we're trying to catchup. This will let
 // system catchup just before the gap.
-TEST_CASE_METHOD(HistoryTests, "too far behind / catchup restart",
-                 "[history][catchupstall]")
+TEST_CASE("too far behind / catchup restart", "[history][catchupstall]")
 {
-    generateAndPublishInitialHistory(1);
+    HistoryTests historyTests{};
+
+    historyTests.generateAndPublishInitialHistory(1);
 
     // Catch up successfully the first time
-    auto app2 =
-        catchupNewApplication(mApp.getLedgerManager().getLastClosedLedgerNum(),
-                              std::numeric_limits<uint32_t>::max(), false,
-                              Config::TESTDB_IN_MEMORY_SQLITE, "app2");
+    auto app2 = historyTests.catchupNewApplication(
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum(),
+        std::numeric_limits<uint32_t>::max(), false,
+        Config::TESTDB_IN_MEMORY_SQLITE, "app2");
 
     // Now generate a little more history
-    generateAndPublishHistory(1);
+    historyTests.generateAndPublishHistory(1);
 
     auto init = app2->getLedgerManager().getLastClosedLedgerNum() + 2;
     REQUIRE(init == 66);
 
     // Now start a catchup on that catchups as far as it can due to gap
     LOG(INFO) << "Starting catchup (with gap) from " << init;
-    REQUIRE(catchupApplication(init, std::numeric_limits<uint32_t>::max(),
-                               false, app2, true, init + 10));
+    REQUIRE(historyTests.catchupApplication(
+        init, std::numeric_limits<uint32_t>::max(), false, app2, true,
+        init + 10));
     REQUIRE(app2->getLedgerManager().getLastClosedLedgerNum() == 75);
 
     app2->getWorkManager().clearChildren();
 
     // Now generate a little more history
-    generateAndPublishHistory(1);
+    historyTests.generateAndPublishHistory(1);
 
     // And catchup successfully
-    init = mApp.getLedgerManager().getLastClosedLedgerNum();
-    REQUIRE(catchupApplication(init, std::numeric_limits<uint32_t>::max(),
-                               false, app2));
+    init = historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
+    REQUIRE(historyTests.catchupApplication(
+        init, std::numeric_limits<uint32_t>::max(), false, app2));
     REQUIRE(app2->getLedgerManager().getLastClosedLedgerNum() == 192);
 }
 
@@ -1278,16 +504,19 @@ TEST_CASE_METHOD(HistoryTests, "too far behind / catchup restart",
  * Test a variety of orderings of CATCHUP_RECENT mode, to shake out boundary
  * cases.
  */
-TEST_CASE_METHOD(HistoryTests, "Catchup recent", "[history][catchuprecent]")
+TEST_CASE("Catchup recent", "[history][catchuprecent]")
 {
+    HistoryTests historyTests{};
+
     auto dbMode = Config::TESTDB_IN_MEMORY_SQLITE;
     std::vector<Application::pointer> apps;
 
-    generateAndPublishInitialHistory(3);
+    historyTests.generateAndPublishInitialHistory(3);
 
     // Network has published 0x3f (63), 0x7f (127) and 0xbf (191)
     // Network is currently sitting on ledger 0xc0 (192)
-    uint32_t initLedger = mApp.getLedgerManager().getLastClosedLedgerNum();
+    uint32_t initLedger =
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
 
     // Check that isolated catchups work at a variety of boundary
     // conditions relative to the size of a checkpoint:
@@ -1298,42 +527,47 @@ TEST_CASE_METHOD(HistoryTests, "Catchup recent", "[history][catchuprecent]")
     for (auto r : recents)
     {
         auto name = std::string("catchup-recent-") + std::to_string(r);
-        apps.push_back(
-            catchupNewApplication(initLedger, r, false, dbMode, name));
+        apps.push_back(historyTests.catchupNewApplication(initLedger, r, false,
+                                                          dbMode, name));
     }
 
     // Now push network along a little bit and see that they can all still
     // catch up properly.
-    generateAndPublishHistory(2);
-    initLedger = mApp.getLedgerManager().getLastClosedLedgerNum();
+    historyTests.generateAndPublishHistory(2);
+    initLedger =
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
 
     for (auto a : apps)
     {
-        catchupApplication(initLedger, 80, false, a);
+        historyTests.catchupApplication(initLedger, 80, false, a);
     }
 
     // Now push network along a _lot_ futher along see that they can all
     // still
     // catch up properly.
-    generateAndPublishHistory(25);
-    initLedger = mApp.getLedgerManager().getLastClosedLedgerNum();
+    historyTests.generateAndPublishHistory(25);
+    initLedger =
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
 
     for (auto a : apps)
     {
-        catchupApplication(initLedger, 80, false, a);
+        historyTests.catchupApplication(initLedger, 80, false, a);
     }
 }
 
 /*
  * Test a variety of LCL/initLedger/count modes.
  */
-TEST_CASE_METHOD(HistoryTests, "Catchup manual", "[history][catchupmanual]")
+TEST_CASE("Catchup manual", "[history][catchupmanual]")
 {
+    HistoryTests historyTests{};
+
     auto dbMode = Config::TESTDB_IN_MEMORY_SQLITE;
     std::vector<Application::pointer> apps;
 
-    generateAndPublishInitialHistory(6);
-    auto initLedger = mApp.getLedgerManager().getLastClosedLedgerNum();
+    historyTests.generateAndPublishInitialHistory(6);
+    auto initLedger =
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
     REQUIRE(initLedger == 383);
 
     for (auto const& test : stellar::gCatchupRangeCases)
@@ -1344,23 +578,24 @@ TEST_CASE_METHOD(HistoryTests, "Catchup manual", "[history][catchupmanual]")
                                 lastClosedLedger, configuration.toLedger(),
                                 configuration.count());
         // manual catchup-recent
-        auto a =
-            catchupNewApplication(configuration.toLedger(),
-                                  configuration.count(), true, dbMode, name);
+        auto a = historyTests.catchupNewApplication(configuration.toLedger(),
+                                                    configuration.count(), true,
+                                                    dbMode, name);
         // manual catchup-complete
-        catchupApplication(initLedger, std::numeric_limits<uint32_t>::max(),
-                           true, a);
+        historyTests.catchupApplication(
+            initLedger, std::numeric_limits<uint32_t>::max(), true, a);
         apps.push_back(a);
     }
 
     // Now push network along a little bit and see that they can all still
     // catch up properly.
-    generateAndPublishHistory(2);
-    initLedger = mApp.getLedgerManager().getLastClosedLedgerNum();
+    historyTests.generateAndPublishHistory(2);
+    initLedger =
+        historyTests.getApp().getLedgerManager().getLastClosedLedgerNum();
 
     for (auto a : apps)
     {
-        catchupApplication(initLedger, 80, false, a);
+        historyTests.catchupApplication(initLedger, 80, false, a);
     }
 }
